pyproject.toml
[project]
name = "ultralytics-mot-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "loguru>=0.7.3",
    "opencv-python-headless>=4.12.0.88",
    "pyyaml>=6.0.0",
    "tqdm>=4.67.1",
    "ultralytics>=8.3.189",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "scikit-learn>=1.0.0",
]

[project.scripts]
mot-infer = "ultralytics_mot_benchmark.cli.infer:main"
mot-params = "ultralytics_mot_benchmark.parameter_testing.parameter_tester:main"
mot-analyze = "ultralytics_mot_benchmark.cli.analyze_filter_logs:main"

[tool.pytest.ini_options]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
